Add extension_redirect endpoint
